stock_trading/pyproject.toml
[project]
name = "stock-trading"
version = "0.1.0"
description = ""
authors = [
    {name = "aqdas1234",email = "bsef21m505@pucit.edu.pk"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "amqp (==5.3.1)",
    "asgiref (==3.9.1)",
    "async-timeout (==5.0.1)",
    "billiard (==4.2.1)",
    "celery (==5.5.3)",
    "click (==8.2.1)",
    "click-didyoumean (==0.3.1)",
    "click-plugins (==1.1.1.2)",
    "click-repl (==0.3.0)",
    "cron-descriptor (==1.4.5)",
    "django (==5.2.4)",
    "django-celery-beat (==2.8.1)",
    "django-environ (==0.12.0)",
    "django-filter (==25.1)",
    "django-timezone-field (==7.1)",
    "djangorestframework (==3.16.0)",
    "djangorestframework-jwt (==1.11.0)",
    "djangorestframework-simplejwt (==5.5.0)",
    "drf-yasg (==1.21.10)",
    "exceptiongroup (==1.3.0)",
    "flower (==2.0.1)",
    "humanize (==4.12.3)",
    "inflection (==0.5.1)",
    "iniconfig (==2.1.0)",
    "kombu (==5.5.4)",
    "packaging (==25.0)",
    "pluggy (==1.6.0)",
    "prometheus-client (==0.22.1)",
    "prompt-toolkit (==3.0.51)",
    "psycopg2-binary (==2.9.10)",
    "pygments (==2.19.2)",
    "pyjwt (==1.7.1)",
    "pytest (==8.4.1)",
    "pytest-django (==4.11.1)",
    "python-crontab (==3.3.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-decouple (==3.8)",
    "pytz (==2025.2)",
    "pyyaml (==6.0.2)",
    "redis (==6.2.0)",
    "six (==1.17.0)",
    "sqlparse (==0.5.3)",
    "tomli (==2.2.1)",
    "tornado (==6.5.1)",
    "typing-extensions (==4.14.1)",
    "tzdata (==2025.2)",
    "uritemplate (==4.2.0)",
    "vine (==5.1.0)",
    "wcwidth (==0.2.13)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
